+++ CHANGELOG.md
@@ -2,13 +2,13 @@
 
 ## 0.2.0
 
-### Added 
+### Added
 
 - Added `--version` flag (#23)
 
 ### Fixed
 
-- Only show diff if it exists (#22) 
+- Only show diff if it exists (#22)
 
 ## 0.1.2