+++ CHANGELOG.md
@@ -2,7 +2,13 @@
 
 ## 0.2.0
 
+### Added
+
 -   Added `--version` flag
+
+### Fixed
+
+-   Only show diff if it exists (#22)
 
 ## 0.1.2