+++ CHANGELOG.md
@@ -1,14 +1,24 @@
 # Changelog
 
+## _Unreleased_
+
+### Added
+-   Added flags `pull-repos`/`--no-pull-repos` to `test` and `run` (#34)
+-   For the `run command`: (#34)
+    -   Show progress for `run`
+    -   Don't sleep after pushing the last PR
+    -   Set a default `push_delay`
+-   Ask for the API key if not passed via parameters
+
+
 ## 0.2.0
 
 ### Added
-
 -   Added `--version` flag (#23)
 
 ### Fixed
-
 -   Only show diff if it exists (#22)
+
 
 ## 0.1.2