+++ CHANGELOG.md
@@ -15,6 +15,7 @@
     -   Ask for the API key if not passed via parameters (#34)
 -   Add `auto-pr progress` command (#35)
 
+
 ## 0.2.0
 
 ### Added